Reorganize pyjinhx into core/, reactive/, and config/ subpackages.
Align internal module layout with usage tiers while keeping the public
from pyjinhx import ... API unchanged; split reactive.py into component
and client modules and document the new structure in README.
