+++ pyjinhx/core/assets.py
