+++ pyjinhx/integrations/redis.py
@@ -13,7 +13,7 @@ import threading
 from collections.abc import Callable
 from typing import Any
 
-from pyjinhx.invalidation import InvalidationBackend
+from pyjinhx.reactive.invalidation import InvalidationBackend
 
 logger = logging.getLogger(__name__)
