+++ pyjinhx/reactive/keys.py
@@ -1,6 +1,6 @@
 from enum import StrEnum
 
-from .utils import ReactiveKey, coerce_load_key, coerce_reactive_keys
+from ..utils import ReactiveKey, coerce_load_key, coerce_reactive_keys
 
 
 class StateKey(StrEnum):
