+++ tests/11_renderer_default_environment_test.py
@@ -2,7 +2,7 @@ import os
 
 from jinja2 import Environment, FileSystemLoader
 
-from pyjinhx.renderer import Renderer
+from pyjinhx.core.renderer import Renderer
 
 
 def test_set_default_environment():
