+++ tests/26_classless_component_assets_test.py
@@ -15,7 +15,7 @@ import tempfile
 from jinja2 import Environment, FileSystemLoader
 
 from pyjinhx import Registry
-from pyjinhx.renderer import Renderer
+from pyjinhx.core.renderer import Renderer
 
 
 def test_classless_component_js_collected():
