+++ tests/33_client_runtime_test.py
@@ -1,6 +1,6 @@
 from markupsafe import Markup
 
-from pyjinhx.reactive import PJX_MOUNTED_HEADER, client_script
+from pyjinhx.reactive.client import PJX_MOUNTED_HEADER, client_script
 from pyjinhx.utils import read_client_runtime
